requirements.txt
obs-websocket-py
requests